fix(extension): fixed JSON file error. New README. Update AI model on BTP to gpt-4
fix #107
