+++ .gitignore
@@ -7,4 +7,3 @@ node_modules
 .nyc_output/
 coverage/
 mochawesome-report/
-src/lib/key.json
