--- src/lib/code-actions/aiPromptsAndDocs.json
@@ -1,8 +1,0 @@
-{
-    "other_prompt": "Please provide the code snippets required to access a secret stored on the VALUE_TMP for a EXTENSION_TMP application. Give first the official doc of VALUE_TMP, then how to store the secret with a step by steg guide with a lot of details and finally the code snippet to retrieve the secret in the code. Give your answer in a way to be displayed inside a HTML page between the <body> </body> tag already present.",
-    "btp_prompt": "Please provide the code snippet in EXTENSION_TMP to access to these environement variables: {'VCAP_SERVICES':{'credstore': [{...'credentials': {'password': '','encryption': {'client_private_key': '',...},...,'url': '','username': ''},...}]}}",
-    "btp_documentation": "<h2>BTP Credential Store Step-by-Step Guide:</h2><ol><li><strong>Entitlement Assignment</strong><ul><li>Your application need to be deployed on BTP</li><li>From your global account page, select '<em>Entitlements</em>' -> '<em>Entity Assignments</em>' from the left menu, and choose the target subaccount.</li><li>Select '<em>Edit</em>' -> '<em>Add Service Plans</em>'. Search for '<em>Credential Store</em>' and choose a plan based on these guidelines:<ul><li>For a Consumption-Based Model, we recommend starting with the 'Free' plan.</li><li>If you're on a Subscription-Based Model, only the 'Standard' plan is available (which isn't free).</li><li>For Trial Accounts, opt for the 'Trial' plan.</li></ul></li><li>By default, the authentication type is <a href='https://help.sap.com/docs/credential-store/sap-credential-store/mutual-tls-authentication'>MTLS</a>. You can change for a <a href='https://help.sap.com/docs/credential-store/sap-credential-store/basic-authentication'>Basic Authentication</a> by providing this parameter: <em>{ 'authentication': { 'type': 'basic' } } </em></li></ul> </li><li><strong>Service Binding</strong><ul><li>From your application page, select '<em>Service Bindings</em>' from the left menu, and create a new one for 'Credential Store'.</li></ul></li></ol><p><strong>Note:</strong> If you're on a Consumption-Based Model and want to upgrade from 'Free' to 'Standard':<ul><li>In Cloud Foundry, simply adjust the service settings.</li><li>In Kyma, you'll need to recreate the service instance.</li></ul></p><p>For more information, visit the following links:</p><ul><li>SAP Discovery Center: <a href='https://discovery-center.cloud.sap/serviceCatalog/credential-store?region=all'>Credential Store</a>.</li><li>SAP Help Portal: <a href='https://help.sap.com/docs/credential-store'>SAP Credential Store</a>.</li><li>SAP Community, Technology Blogs: <a href='https://community.sap.com/t5/technology-blogs-by-sap/sap-btp-safely-consume-your-app-secrets-with-credential-store-%EF%B8%8F/ba-p/13534019'>SAP BTP - Safely consume your app secrets with Credential Store 🕵️</a>.</li></ul><p>The 'VCAP_SERVICES' environment variable stores essential REST API endpoint information for the SAP credential store, alongside the associated credentials required to access the API. Additionally, it includes details of the 'client_private_key,' utilized for decrypting the response payload. The structure of this variable may vary slightly depending on the chosen authentication method. Here an example of VCAP_SERVICES with the Basic Auth type and all the information you will need to retrieve and decrypt the payload: </p><pre> <code> {'VCAP_SERVICES':{'credstore': [{...'credentials': {'password': '','encryption': {'client_private_key': '',...},...,'url': '','username': ''},...}]}}</code> </pre><p><strong>And here is some code (generated by an AI) on how to access 'VCAP_SERVICES', based on your file extension:</strong></p><pre><code>CODE_TMP</code></pre><p> The HTTP header should include a custom namespace header labeled 'sapcp-credstore-namespace'. </p><p> The payload encryption is based on JWE compact serialization format. Check JSON Web Encryption.<ul>Supported algorithms:<ul><li>Key encryption: RSA-OAEP-256</li><li>Content encryption: A256GCM </li></ul></ul>You will need the private key to decrypt it.</p>",
-    "begin_html": "<!DOCTYPE html> <html lang='en'> <head> <meta charset='UTF-8'> <meta name='viewport' content='width=device-width, initial-scale=1.0'> <title>VALUE_TMP Step-by-Step Guide</title> <style> body { font-family: Arial, sans-serif; margin: 20px; line-height: 1.6; /* Adjust line height */ } h2 { color: #007acc; } ol { margin-left: 20px; } </style> </head> <body>",
-    "end_html": "</body> </html>",
-    "promt_env_var": "Could you please provide me all the steps to follow to store my secret in a .env file? Can you make your reponse in one comment block for a EXTENSION_TMP file."
-}
