+++ src/lib/code-actions/openai-btp.ts
@@ -3,6 +3,8 @@ import { Uri } from 'vscode';
 
 import axios, { AxiosResponse } from 'axios';
 
+import LoggerFactory from '../logger-factory';
+
 interface SvcKey {
     url: string;
     uaa: {
@@ -14,10 +16,11 @@ interface SvcKey {
 
 interface Data {
     deployment_id: string;
-    prompt: string;
+    messages: { role: string; content: string }[];
     max_tokens: number;
     temperature: number;
-    n: number;
+    frequency_penalty: number;
+    presence_penalty: number;
 }
 
 async function loadUserKey(keyPath: string): Promise<SvcKey> {
@@ -62,7 +65,9 @@ async function makeRequest(
 
         return response;
     } catch (error) {
-        console.error('Error generating chat response:', error);
+        LoggerFactory.getInstance().error('Error generating chat response:', {
+            error: error,
+        });
         return 'Error generating chat response: ' + error;
     }
 }
@@ -70,22 +75,34 @@ async function makeRequest(
 export async function call_openAI_btp(message: string, keyPath: string) {
     try {
         const data: Data = {
-            deployment_id: 'text-davinci-003',
-            prompt: message,
-            max_tokens: 1000,
-            temperature: 0.0,
-            n: 1,
+            deployment_id: 'gpt-4',
+            messages: [{ role: 'user', content: message }],
+            max_tokens: 800,
+            temperature: 1,
+            frequency_penalty: 0,
+            presence_penalty: 0,
         };
+
         const response = (await makeRequest(
             data,
             keyPath,
         )) as unknown as AxiosResponse;
-        return { success: true, message: response.data.choices[0].text };
+
+        if (response.data === undefined) {
+            throw new Error(response.toString());
+        }
+
+        return {
+            success: true,
+            message: response.data.choices[0].message.content,
+        };
     } catch (error) {
-        console.error('Error generating chat response: ', error);
+        LoggerFactory.getInstance().error('Error generating chat response:', {
+            error: error,
+        });
         return {
             success: false,
-            message: 'Error generating chat response: ' + error,
+            message: error,
         };
     }
 }
